+++ web/src/views/ChatView.vue
@@ -994,6 +994,7 @@
         <ChatEditor
           ref="chatEditorRef"
           :placeholder="activeSpace ? `Message ${activeSpace.name}...` : undefined"
+          :member-names="mentionMemberNames"
           @send="handleEditorSend"
         />
       </div>
@@ -1104,6 +1105,7 @@ import type { Ref } from 'vue'
 import { useRouter } from 'vue-router'
 import { useSpaceTimeline } from '../composables/useSpaceTimeline'
 import { ChatEditor } from '../components/ChatEditor'
+import { spaceRosterNames } from '../components/ChatEditor/mentionSuggestions'
 import { ThreadPanel } from '../components/ThreadPanel'
 import SwarmStatus from '../components/SwarmStatus.vue'
 import ThreadDetail from '../components/ThreadDetail.vue'
@@ -1210,6 +1212,7 @@ watch(() => props.spaceId, async (newId) => {
 const { sessions, getMessages, fetchMessages, queueIfHydrating, formatSessionLabel, renameSession,
   getLastSeenMessageId, setLastSeenMessageId } = useSessions()
 const { activeSpace, dms, openDM } = useSpaces()
+const mentionMemberNames = computed(() => spaceRosterNames(activeSpace.value))
 
 function isKnownSession(id: string): boolean {
   return sessions.value.some(s => s.id === id)